enhancement(rafty): Improve election process
